Rename spriteIndex parameter and args
I saw the method in SpriteSheet class was called on uiBtnIconSheet passing either 0, 1 or 2 as parameter, that spritesheet has 3 elements so I guessed the method parameter is the index of the element in the spritesheet. Soearching all the calls to this method, I also renamed all variables used as argument.