Rename method setCurrentIndex
I had to use a generic name because if an animation table is added to the spritesheet then the meaning of the method goes from "set sprite index" to "set frame index". Maybe frameIndex would be the better choce, since spriteIndex does not apply here since the anim table can output a sprite with a different index.