fix: `xp` and `mapPixelX` merged due to decompilation artifact
This in an important change because it could be a game breaking error
that I did not fully understand before.

The mess happened because Unit inherits from Spritesheet, and
both classes had a field with the same type and name, and VSCode
did not complain about that, and I did not notice. After using
"rename all references" on the field in the Unit class
(now named mapPixelX), VSCode renamed just a portion of all actual
referenced, the ones that could trace to Unit class, while the true
"mapPixelX" is the one on the SpriteSheet class.
Another field "l" which is basically a not yet renamed "mapPixelY"
is correctly casted to the code is clear that field comes fron the
parent SpriteSheet class.

Luckily as of now I think I have caught and fixed all or most of these
"same name" issues, but holy crap I thought tooling would prevent them.
To avoid huge mistakes one needs to be very careful and that's why I'm
keeping commits very small and easy to review.

After a quick check, both the CFR and Fernflower decompilators have
this issue with parent and child classes resulting in fields with
same name and type, which is probably a Java issue, or a VSCode one.
What kind of language allows to shadow fields of the inherited class
and leave this ambiguity?