refactor: rename and document `tryLevelUp` method
`75 << this.stars` is a pretty nice trick, I've just learned
that shifting left doubles a number. This could also be some
automatic compiler optimization, I guess one needs to be careful
about number representation when bit shifting (maybe not for JVM since
Java runs on that and not directly on system).