refactor: explain counterintuitive condition
The way this condition is writte is unreadable, I had to look at this
for a bit to undestand. I suspect this is a compiler optimization since
these three "if" wwre originally combined in a unique stupidly long one.
Maybe the compiler used one of the logical equivalence laws to
converti this from readable to unreadable, or it's the decompiler
that had trouble making it readable. Not going to dig deeper,
it's just very inconvenient.