refactor: remove useless check in condition
There was just something fishy about this condition, I could not
make sense of it: (state == -1) then the second part is irrelevant
(-1 is always different from 3), if it's false then the whole
expression is false without needing to check the second operand. WTF?
I thought I was stupid for not getting what this code does, but
after double-checking with a spreadsheet the truth table for the
original condition and mune, it turns out I am right - that second
check is completely useless, and somebody shipped a game
with a method that works by chance.
That said, it's relieving to see how messy this whole if chain was
before deobfuscation, and look at it now!