Rename field "ssUnitsBattleAnimations"
Notice how the spritesheet for a specific unit is created only if
needed (the null check). Compared to the usual approach of loading
everything at game start, it's a nice technique to save memory
at the cost of some performance at first load. Or maybe it's to avoid
loading them all at once. A sort of memoization, it's interesting how
even for such a small game some tradeoffs were necessary
for mobile devices.